Rename block_processing.rs -> block_verification
