+++ beacon_node/beacon_chain/src/block_verification.rs
