+++ beacon_node/beacon_chain/src/lib.rs
@@ -4,7 +4,7 @@ extern crate lazy_static;
 
 mod beacon_chain;
 mod beacon_snapshot;
-mod block_processing;
+mod block_verification;
 pub mod builder;
 mod errors;
 pub mod eth1_chain;
@@ -26,7 +26,7 @@ pub use self::beacon_chain::{
 };
 pub use self::beacon_snapshot::BeaconSnapshot;
 pub use self::errors::{BeaconChainError, BlockProductionError};
-pub use block_processing::BlockError;
+pub use block_verification::BlockError;
 pub use eth1_chain::{Eth1Chain, Eth1ChainBackend};
 pub use events::EventHandler;
 pub use fork_choice::ForkChoice;
